# app/api/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.ChannelType:
    enum:
    - email
    - sms
    - push
    - webhook
    type: string
    x-enum-varnames:
    - ChannelEmail
    - ChannelSMS
    - ChannelPush
    - ChannelWebhook
  domain.Group:
    properties:
      description:
        type: string
      groupId:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      orgId:
        type: string
      subscribers:
        items:
          type: string
        type: array
    type: object
  domain.ListResult-domain_Group:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.Group'
        type: array
      pagination:
        $ref: '#/definitions/domain.PaginationResult'
    type: object
  domain.ListResult-domain_Rule:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.Rule'
        type: array
      pagination:
        $ref: '#/definitions/domain.PaginationResult'
    type: object
...
  domain.PaginationResult:
    properties:
      page:
        type: integer
      pageSize:
        type: integer
      totalCount:
        type: integer
      totalPages:
        type: integer
    type: object
  domain.Recipients:
    properties:
      broadcast:
        type: boolean
      groups:
        items:
          type: string
        type: array
      subscriberIds:
        items:
          type: string
        type: array
    type: object
  domain.Rule:
    properties:
      channels:
        items:
          $ref: '#/definitions/domain.ChannelType'
        type: array
      createdAt:
        type: string
      defaultRecipients:
        $ref: '#/definitions/domain.Recipients'
      eventType:
        type: string
      orgId:
        type: string
      templateRefs:
        additionalProperties:
          type: string
        type: object
      updatedAt:
        type: string
    type: object
...
  internal_handlers_group.groupRequest:
    properties:
      description:
        type: string
      groupId:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      subscribers:
        items:
          type: string
        type: array
    type: object
  internal_handlers_rule.ruleRequest:
    properties:
      channels:
        items:
          $ref: '#/definitions/domain.ChannelType'
        type: array
      defaultRecipients:
        $ref: '#/definitions/domain.Recipients'
      eventType:
        type: string
      templateRefs:
        additionalProperties:
          type: string
        type: object
    type: object
...
info:
  contact: {}
  description: API for templates, groups, subscribers, and rules
  title: NotifyX API
  version: "1.0"
paths:
  /groups:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Sort field
        in: query
        name: sortBy
        type: string
      - default: asc
        description: Sort order (asc/desc)
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ListResult-domain_Group'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List groups with pagination
      tags:
      - groups
    post:
      consumes:
      - application/json
      parameters:
      - description: Group data
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_group.groupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Group'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new group
      tags:
      - groups
  /groups/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a group
      tags:
      - groups
    get:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Group'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a group by ID
      tags:
      - groups
    put:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Group patch data
        in: body
        name: group
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Group'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a group (merge patch)
      tags:
      - groups
  /rules:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      - description: Sort field
        in: query
        name: sortBy
        type: string
      - default: asc
        description: Sort order (asc/desc)
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ListResult-domain_Rule'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List rules with pagination
      tags:
      - rules
    post:
      consumes:
      - application/json
      parameters:
      - description: Rule data
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_rule.ruleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Rule'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new rule
      tags:
      - rules
  /rules/{eventType}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Event Type
        in: path
        name: eventType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a rule
      tags:
      - rules
    get:
      consumes:
      - application/json
      parameters:
      - description: Event Type
        in: path
        name: eventType
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Rule'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a rule by event type
      tags:
      - rules
    put:
      consumes:
      - application/json
      parameters:
      - description: Event Type
        in: path
        name: eventType
        required: true
        type: string
      - description: Rule patch data
        in: body
        name: rule
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Rule'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a rule (merge patch)
      tags:
      - rules
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"